Split widget: effects timeline
